refactor: move business-related enums from infrastructure to domain
