+++ src/main/kotlin/com/dobby/backend/domain/enums/MatchType.kt
@@ -1,0 +1,5 @@
+package com.dobby.backend.domain.enums
+
+enum class MatchType {
+    OFFLINE, ONLINE, ALL
+}
