+++ src/main/kotlin/com/dobby/backend/domain/enums/VerificationStatus.kt
@@ -1,0 +1,6 @@
+package com.dobby.backend.domain.enums
+
+enum class VerificationStatus {
+    HOLD,
+    VERIFIED
+}
