+++ src/main/kotlin/com/dobby/backend/domain/enums/member/GenderType.kt
@@ -1,0 +1,5 @@
+package com.dobby.backend.domain.enums.member
+
+enum class GenderType {
+    MALE, FEMALE, ALL
+}
