+++ src/main/kotlin/com/dobby/backend/domain/enums/member/MemberStatus.kt
@@ -1,0 +1,6 @@
+package com.dobby.backend.domain.enums.member
+
+enum class MemberStatus {
+    HOLD,
+    ACTIVE
+}
