+++ src/main/kotlin/com/dobby/backend/domain/enums/member/ProviderType.kt
@@ -1,0 +1,5 @@
+package com.dobby.backend.domain.enums.member
+
+enum class ProviderType {
+    NAVER, GOOGLE
+}
