+++ src/main/kotlin/com/dobby/backend/domain/enums/member/RoleType.kt
@@ -1,4 +1,4 @@
-package com.dobby.backend.infrastructure.database.entity.enums.member
+package com.dobby.backend.domain.enums.member
 
 enum class RoleType(
     val roleName: String
