--- src/main/kotlin/com/dobby/backend/infrastructure/database/entity/enums/MatchType.kt
@@ -1,5 +1,0 @@
-package com.dobby.backend.infrastructure.database.entity.enums
-
-enum class MatchType {
-    OFFLINE, ONLINE, ALL
-}
