--- src/main/kotlin/com/dobby/backend/infrastructure/database/entity/enums/VerificationStatus.kt
@@ -1,6 +1,0 @@
-package com.dobby.backend.infrastructure.database.entity.enums
-
-enum class VerificationStatus {
-    HOLD,
-    VERIFIED
-}
