--- src/main/kotlin/com/dobby/backend/infrastructure/database/entity/enums/member/GenderType.kt
@@ -1,5 +1,0 @@
-package com.dobby.backend.infrastructure.database.entity.enums.member
-
-enum class GenderType {
-    MALE, FEMALE, ALL
-}
