--- src/main/kotlin/com/dobby/backend/infrastructure/database/entity/enums/member/MemberStatus.kt
@@ -1,6 +1,0 @@
-package com.dobby.backend.infrastructure.database.entity.enums.member
-
-enum class MemberStatus {
-    HOLD,
-    ACTIVE
-}
