--- src/main/kotlin/com/dobby/backend/infrastructure/database/entity/enums/member/ProviderType.kt
@@ -1,5 +1,0 @@
-package com.dobby.backend.infrastructure.database.entity.enums.member
-
-enum class ProviderType {
-    NAVER, GOOGLE
-}
